requirements.txt
pyyaml>=6.0
boto3
openai
pytest
streamlit>=1.0